dom/indexedDB/test/test_filehandle_progress_events.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function* testSteps()
  {
    const name = window.location.pathname;

    var testBuffer = getRandomBuffer(100000);

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    db.onerror = errorHandler;

    request = db.createMutableFile("test.txt");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    let fileHandle = mutableFile.open("readwrite");

    let sum = 0;

    request = fileHandle.write(testBuffer);
    request.onprogress = function(event) {
      let loaded = event.loaded;
      let total = event.total;
      ok(loaded >= 0 && loaded <= total, "Correct loaded progress");
      is(total, testBuffer.byteLength, "Correct total progress");
      sum += event.loaded - sum;
    }
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    is(sum, testBuffer.byteLength, "Correct loaded progress sum");

    sum = 0;

    fileHandle.location = 0;
    request = fileHandle.readAsArrayBuffer(testBuffer.byteLength);
    request.onprogress = function(event) {
      let loaded = event.loaded;
      let total = event.total;
      ok(loaded >= 0 && loaded <= total, "Correct loaded progress");
      is(total, testBuffer.byteLength, "Correct total progress");
      sum += event.loaded - sum;
    }
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    is(sum, testBuffer.byteLength, "Correct loaded progress sum");

    finishTest();
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
